♻️ refactor: move providers pages to dashboard subdirectory
Relocate all provider-related pages, components, and tests from
/providers to /dashboard/providers for better route organization
