+++ src/app/(dashboard)/dashboard/providers/[providerId]/page.tsx
