+++ src/app/(dashboard)/dashboard/providers/[providerId]/provider-charts-section.tsx
